pyproject.toml
[project]
name = "uber-hacksaw"
version = "0.1.0"
description = "Malware Scanner written in Python"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
authors = [
    { name="Eric Harris", email="ericktheredd5875@gmail.com" }
]
dependencies = [
    "entropy>=0.1.5",
    "loguru>=0.7.3",
    "pefile>=2024.8.26",
    "ppdeep>=20250625",
    "rich>=14.1.0",
    "typer>=0.17.3",
    "yara-python>=4.5.0",
]

[project.scripts]
uber-hacksaw = "uber_hacksaw.cli:app"

[project.urls]
Homepage = "https://github.com/ericktheredd5875/uber-hacksaw"

[dependency-groups]
dev = [
    "bandit>=1.8.6",
    "black>=25.1.0",
    "psutil>=7.1.0",
    "pyright>=1.1.404",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "pytest-pretty>=1.3.0",
    "ruff>=0.12.11",
    "tree>=0.2.4",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
addopts = "-q"
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
src = ["src"]
target-version = "py313"
lint.select = ["E","F","I","UP","B"]

lint.ignore = [
    "E501",  # line length handled by formatter
]
exclude = [
    "**/tests",
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = [
    "S101", # Use of `assert` detected
    "D103", # Missing docstring in public function
]
#C:\CodeBases\uber-hacksaw\src\uber_hacksaw\static\type_id.py
"**/type_id.py" = [
    "F401", # Imported but unused
]
"docs/**" = [
    "INP001",   # Requires __init__.py but docs folder is not a package.
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.bandit]
exclude_dirs = ["tests", ".venv"]
#tests = ["B201", "B301"]
#skips = ["B101", "B601"]